requirements.txt
flask
torch
torchvision
transformers
peft
pillow
kagglehub
opencv-python
numpy
